Converted frequency to long primitive type (from integer) in anticipation of HackRF support which requires frequencies greater than 2.147 GHz, which is the max hertz value that can be represented by an integer primitive.

+++ src/controller/channel/ChannelMap.java
@@ -94,7 +94,7 @@ public class ChannelMap
 		mName = name;
 	}
 	
-	public int getFrequency( int channelNumber )
+	public long getFrequency( int channelNumber )
 	{
 		for( ChannelRange range: mRanges )
 		{
