+++ src/decode/fleetsync2/FleetsyncCallEvent.java
@@ -63,7 +63,7 @@ public class FleetsyncCallEvent extends CallEvent
     }
 
     @Override
-    public int getFrequency()
+    public long getFrequency()
     {
         return 0;
     }
